mvproject/requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
filelock==3.9.0
platformdirs==2.6.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
virtualenv==20.17.1
